pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "npa"
version = "0.1.0"
description = "Next Policy Agent — A high-performance, secure policy engine (Python/FastAPI rewrite of OPA)"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
authors = [
    {name = "NPA Team"},
]
keywords = ["policy", "authorization", "rego", "opa", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",

    # Performance
    "orjson>=3.10.0",
    "uvloop>=0.21.0; sys_platform != 'win32'",

    # Security / Crypto
    "cryptography>=44.0.0",
    "PyJWT>=2.9.0",

    # Configuration
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",

    # Storage
    "aiosqlite>=0.20.0",

    # CLI
    "typer>=0.15.0",
    "rich>=13.9.0",

    # Observability
    "prometheus-client>=0.21.0",
    "structlog>=24.4.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",

    # Networking
    "httpx>=0.28.0",

    # Utilities
    "cachetools>=5.5.0",
    "xxhash>=3.5.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "httpx>=0.28.0",
]

[project.scripts]
npa = "npa.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["npa"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "S", "B", "A", "C4", "UP", "RUF"]
ignore = ["S101"]  # allow assert in tests

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
